@@ -1,8 +1,10 @@
 class Answer < ActiveRecord::Base
-  attr_accessible :value, :field_name, :business_id
+  #attr_accessible :value, :field_name, :business_id
 
   belongs_to :business
   has_one :user, :through => :business
 
   validates_presence_of :value
+
+  private 
 end
